refactor: extract shared FIT to SSI tool UI
