+++ src/components/tools/FitToSsiDiveLogHelper/assets/ssi-app-showing-the-qr-code-scanner.png
