+++ src/test/mocks/ideal-image.tsx
@@ -6,7 +6,6 @@ interface Props {
 
 /** Lightweight stand-in for the Docusaurus IdealImage wrapper in tests. */
 const IdealImage = ({ alt = '', width, height }: Props) => (
-  // eslint-disable-next-line @next/next/no-img-element
   <img alt={alt} width={width} height={height} />
 )
